Latest tests

+++ database/migrations/2021_05_07_234622_add_tickrate_column.php
@@ -14,7 +14,7 @@ class AddTickrateColumn extends Migration
     public function up()
     {
         Schema::table('PlayerTiming', function (Blueprint $table) {
-            $table->unsignedInteger('Tickrate')->after('ZoneOrdinal');
+            $table->double('Tickrate')->after('ZoneOrdinal');
         });
     }
